Refactor: refator to match review suggestions
